update the c++ api
