--- tests/src/test_mobilenet.cpp
@@ -1,232 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * License); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * AS IS BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-/*
- * Copyright (c) 2017, Open AI Lab
- * Author: haitao@openailab.com
- */
-#include <unistd.h>
-#include <sys/time.h>
-
-#include <iostream>
-#include <functional>
-#include <algorithm>
-#include <fstream>
-#include <iomanip>
-
-#include "tengine_operations.h"
-#include "tengine_c_api.h"
-
-const char* model_file = "./models/mobilenet.tmfile";
-const char* image_file = "./tests/images/cat.jpg";
-const char* label_file = "./models/synset_words.txt";
-
-const float channel_mean[3] = {104.007, 116.669, 122.679};
-
-int repeat_count = 100;
-
-unsigned long get_cur_time(void)
-{
-    struct timeval tv;
-
-    gettimeofday(&tv, NULL);
-
-    return (tv.tv_sec * 1000000 + tv.tv_usec);
-}
-
-void LoadLabelFile(std::vector<std::string>& result, const char* fname)
-{
-    std::ifstream labels(fname);
-
-    std::string line;
-    while(std::getline(labels, line))
-        result.push_back(line);
-}
-
-static inline bool PairCompare(const std::pair<float, int>& lhs, const std::pair<float, int>& rhs)
-{
-    return lhs.first > rhs.first;
-}
-
-static inline std::vector<int> Argmax(const std::vector<float>& v, int N)
-{
-    std::vector<std::pair<float, int>> pairs;
-    for(size_t i = 0; i < v.size(); ++i)
-        pairs.push_back(std::make_pair(v[i], i));
-    std::partial_sort(pairs.begin(), pairs.begin() + N, pairs.end(), PairCompare);
-
-    std::vector<int> result;
-    for(int i = 0; i < N; ++i)
-        result.push_back(pairs[i].second);
-    return result;
-}
-
-void get_input_data(const char* image_file, float* input_data, int img_h, int img_w, const float* mean, float scale)
-{
-    image img = imread(image_file);
-
-    image resImg = resize_image(img, img_w, img_h);
-    resImg = rgb2bgr_premute(resImg);
-    float* img_data = ( float* )resImg.data;
-    int hw = img_h * img_w;
-    for(int c = 0; c < 3; c++)
-        for(int h = 0; h < img_h; h++)
-            for(int w = 0; w < img_w; w++)
-            {
-                input_data[c * hw + h * img_w + w] = (*img_data - mean[c]) * scale;
-                img_data++;
-            }
-}
-
-int main(int argc, char* argv[])
-{
-    int res;
-
-    while((res = getopt(argc, argv, "r:")) != -1)
-    {
-        switch(res)
-        {
-            case 'r':
-                repeat_count = strtoul(optarg, NULL, 10);
-                break;
-
-            default:
-                break;
-        }
-    }
-
-    int img_h = 224;
-    int img_w = 224;
-
-    /* prepare input data */
-    float* input_data = ( float* )malloc(sizeof(float) * img_h * img_w * 3);
-
-    get_input_data(image_file, input_data, img_h, img_w, channel_mean, 0.017);
-
-    for(int i = 0; i < 10; ++i)
-        printf("%f\n", input_data[i]);
-
-    init_tengine();
-
-    std::cout << "run-time library version: " << get_tengine_version() << "\n";
-
-    if(request_tengine_version("1.0") < 0)
-        return -1;
-
-    graph_t graph = create_graph(nullptr, "tengine", model_file);
-
-    if(graph == nullptr)
-    {
-        std::cout << "Create graph0 failed\n";
-        std::cout << "errno: " << get_tengine_errno() << "\n";
-        return -1;
-    }
-
-    /* get input tensor */
-    int node_idx = 0;
-    int tensor_idx = 0;
-
-    tensor_t input_tensor = get_graph_input_tensor(graph, node_idx, tensor_idx);
-
-    if(input_tensor == nullptr)
-    {
-        std::printf("Cannot find input tensor,node_idx: %d,tensor_idx: %d\n", node_idx, tensor_idx);
-        return -1;
-    }
-
-    int dims[] = {1, 3, img_h, img_w};
-
-    set_tensor_shape(input_tensor, dims, 4);
-
-    /* setup input buffer */
-
-    if(set_tensor_buffer(input_tensor, input_data, 3 * img_h * img_w * 4) < 0)
-    {
-        std::printf("Set buffer for tensor failed\n");
-        return -1;
-    }
-
-    /* run the graph */
-    int ret_prerun = prerun_graph(graph);
-    if(ret_prerun < 0)
-    {
-        std::printf("prerun failed\n");
-        return -1;
-    }
-
-    dump_graph(graph);
-
-    run_graph(graph, 1);
-
-    // benchmark start here
-    printf("REPEAT COUNT= %d\n", repeat_count);
-
-    unsigned long start_time = get_cur_time();
-
-    for(int i = 0; i < repeat_count; i++)
-        run_graph(graph, 1);
-
-    unsigned long end_time = get_cur_time();
-
-    unsigned long off_time = end_time - start_time;
-    std::printf("Repeat [%d] time %.2f us per RUN. used %lu us\n", repeat_count, 1.0f * off_time / repeat_count,
-                off_time);
-
-    /* get output tensor */
-    tensor_t output_tensor = get_graph_output_tensor(graph, node_idx, tensor_idx);
-
-    if(output_tensor == nullptr)
-    {
-        std::printf("Cannot find output tensor , node_idx: %d,tensor_idx: %d\n", node_idx, tensor_idx);
-        return -1;
-    }
-
-    int count = get_tensor_buffer_size(output_tensor) / 4;
-
-    float* data = ( float* )(get_tensor_buffer(output_tensor));
-    float* end = data + count;
-
-    std::vector<float> result(data, end);
-
-    std::vector<int> top_N = Argmax(result, 5);
-
-    std::vector<std::string> labels;
-
-    LoadLabelFile(labels, label_file);
-
-    for(unsigned int i = 0; i < top_N.size(); i++)
-    {
-        int idx = top_N[i];
-
-        std::cout << std::fixed << std::setprecision(4) << result[idx] << " - \"";
-        std::cout << labels[idx] << "\"\n";
-    }
-
-    release_graph_tensor(output_tensor);
-    release_graph_tensor(input_tensor);
-    postrun_graph(graph);
-    destroy_graph(graph);
-
-    free(input_data);
-
-    std::cout << "ALL TEST DONE\n";
-
-    release_tengine();
-    return 0;
-}
