--- tests/src/vgg16_mem.sh
@@ -1,6 +1,0 @@
-#!/bin/bash 
-
-export REPEAT_COUNT=10
-export LOW_MEM_MODE=1
-
-./build/tests/bin/vgg16 ./models/ ./tests/images/bike.jpg
