--- tools/default.config
@@ -1,17 +1,0 @@
-#The following option is used for tengine root path
-TENGINE_ROOT=/home/qwang/OAL/test/tengine
-#The following option is used for system architecture type [x86/Arm64/Arm32]
-ARCH_TYPE="Arm64"
-#The following option is used for android ndk path
-ANDROID_NDK=/android_ndk/path
-#The following option is used for android api level
-API_LEVEL=22
-#The following option is used for setting cross compiler path
-#EMBEDDED_CROSS_ROOT=/opt/hisi-linux/x86-arm/arm-himix200-linux/bin/
-#The following option is used for setting cross compiler prefix
-#CROSS_COMPILE=arm-himix200-linux-
-#The following two options are used for setting protobuf
-PROTOBUF_INCLUDE_PATH=/protobuf/path/include
-PROTOBUF_LIB_PATH=/protobuf/path/lib
-#The following option are used for setting acl
-ACL_ROOT=/acl/path
